requirements.txt
Flask
Pillow==11.1.0
pillow-heif==0.21.0
Flask-SQLAlchemy==3.1.1
cryptography
requests
Flask-Migrate
ollama
httpx
APScheduler>=3.9.0
openai
celery
redis
open_clip_torch
scikit-learn
gunicorn
psutil>=5.9.0  # For memory monitoring